refactor: move external values to runtime_structure::external_values
